Renaming plugins to scaffolders as well as minor updates to the code.
